@@ -15,25 +16,15 @@
 	<version>0.0.7-SNAPSHOT</version>
 
 	<properties>
-		<spring.session.jdbc.version>2.7.0</spring.session.jdbc.version>
+		<java.version>17</java.version>
 
 		<maven.javadoc.failOnError>false</maven.javadoc.failOnError>
-		<jdk.version>1.8</jdk.version>
-		<modelmapper.version>3.1.1</modelmapper.version>
-		<testng.version>7.5.1</testng.version>
-		<findbugs.maven.plugin.version>3.0.5</findbugs.maven.plugin.version>
-		<jacoco.maven.plugin.version>0.8.8</jacoco.maven.plugin.version>
-		<h2.version>1.4.200</h2.version>
-		<open-api.version>1.6.7</open-api.version>
-
-		<!--Sonar rule configuration-->
-		<sonar.issue.ignore.multicriteria>e1,e2</sonar.issue.ignore.multicriteria>
-		<!--Suppressed "Utility class should not have a public constructor"-->
-		<sonar.issue.ignore.multicriteria.e1.ruleKey>squid:S1118</sonar.issue.ignore.multicriteria.e1.ruleKey>
-		<sonar.issue.ignore.multicriteria.e1.resourceKey>**/*SpringBootTomcatApplication.java</sonar.issue.ignore.multicriteria.e1.resourceKey>
-		<!--Suppressed "Resources should be closed" warning for Spring context-->
-		<sonar.issue.ignore.multicriteria.e2.ruleKey>squid:S2095</sonar.issue.ignore.multicriteria.e2.ruleKey>
-		<sonar.issue.ignore.multicriteria.e2.resourceKey>**/*SpringBootTomcatApplication.java</sonar.issue.ignore.multicriteria.e2.resourceKey>
+		<modelmapper.version>3.2.0</modelmapper.version>
+		<testng.version>7.8.0</testng.version>
+		<mockito.version>5.2.0</mockito.version>
+		<open-api.version>1.7.0</open-api.version>
+		<maven-clean-plugin.version>3.3.2</maven-clean-plugin.version>
+		<maven-compiler-plugin.version>3.11.0</maven-compiler-plugin.version>
 	</properties>
 
 	<dependencies>
@@ -181,24 +144,12 @@
 				<artifactId>spring-boot-maven-plugin</artifactId>
 			</plugin>
 			<plugin>
-				<groupId>org.jacoco</groupId>
-				<artifactId>jacoco-maven-plugin</artifactId>
-				<version>${jacoco.maven.plugin.version}</version>
-				<executions>
-					<execution>
-						<id>pre-unit-test</id>
-						<goals>
-							<goal>prepare-agent</goal>
-						</goals>
-					</execution>
-					<execution>
-						<id>post-unit-test</id>
-						<phase>test</phase>
-						<goals>
-							<goal>report</goal>
-						</goals>
-					</execution>
-				</executions>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>${maven-compiler-plugin.version}</version>
+				<configuration>
+					<source>17</source>
+					<target>17</target>
+				</configuration>
 			</plugin>
 			<plugin>
 				<artifactId>maven-clean-plugin</artifactId>
